[zimwriterfs] Try to avoid too big cluster.
We check that cluster will not be too big *before* adding the content.
This way, cluster are always closed before the maximum size and not
just after.

The only way a cluster can be too big is if the content of a sole article
is bigger than the maximum size.

Change-Id: I77a581df46ae87e01a3fe2689570a7c7355d1877