Provide time gap between uuid generation during tests
GNU Mach's i.e. Hurd's kernel clock is not very accurate so during
test on generating and comparing uuid's time might be same leading to
generating same Uuid. This patch adds a sleep of 1s between 2
generate statemens. Thanks to Pino Toscano.

This patch was written for the Debian package to fix a build failure on
GNU Hurd, and is suitable to be included upstream.

Change-Id: I1a1520bb0597244a1d2423ed17e87c5a39cb4aa6