Add travis compilation.
We are using kiwix-build tool from kiwix organization to prepare the
dependencies of libzim. The dependencies are also store on kiwix public
servers.

This is not technically necessary but it simplifies a lot our process.

Ideally we should not depends on kiwix tools (and more especially kiwix
infrastucture) but we have very few options for now. The other options
would have be to recompile all libzim dependencies by hand in travis
scripts (for all architectures).

Maybe kiwix teams will make kiwix-build a separate project and we will
be able to use the tool without depending of kiwix.