Remove clusterImpl.
No need to have a `ClusterImpl` class, the same behaviour can be achieved
with std::shared_ptr.

Using std::shared_ptr simplify the code and it is probably a bit more
efficient as it uses template and avoid intermediate function calls.

As `Cluster` class is an internal class and should not be used by user
code, we can allow us to change it.
(But `Blob` need to be changed, so it is a API change)