Move to buffer strategy.
This is a big change in the way libzim is working.

Before that zim reading was axed around a file stream who was
acting as a reader. This reader has to be placed (seek) and the data read from
it each time we need to get some content. It induces some memory copy between
the file and internal buffer and (most of the time) between final buffer.

With this commit, we are moving to a reader/buffer oriented way of reading.

Readers are lightweigth object just representing a range than can be read.
You can get a sub reader from it (a reader with a range inside the "parent"
one) or a buffer from it (from same or small range)

Buffer are actual memory buffer. You can get the adresse of the memory buffer,
the size of the buffer. You can also get a sub buffer from it. It is a buffer
sharing the same data of the "parent" buffer.

As most of the time we need to get range of data from the zim, the buffer
strategy is closest from what the reading will be.

On top of that, we can make some optimisation later (mmap, ...)