Correctly stem term for the research for "old" zims.
Zim files created before 2017-03 doesn't have a language metadata in the
database.

But the database was created stemming the content. So we need to stem
our request also.

For example, "Maladie" will be indexed using the stemmed term "malad".
If we don't stem our request, xapian will not found the "maladie" term
in the database, and will return no result.
If we stem our request, we search for "malad" and everything works.

So we need a language. zim file itself store a language metadata, so
let's use it if the database has no metadata.