Use the correct type to store size/offset.
As a zim files may be larger than 4GB, we need 8 bytes len integer
to store size or offset in a zim file.

`std::size_t` is not the same depending if the architecture is
32 bits (4 bytes) or 64 bits (8 bytes).

On the other side, we are properly declaring `offset_type` in zim.h to
always be 8 bytes whatever the architecture is.

So, let's use `offset_type` instead of `std::size_t`.

Fixes kiwix/kiwix-tools#76