Rewrite how we handle endianness.
As said here:
https://commandcenter.blogspot.fr/2012/04/byte-order-fallacy.html

We should not care about the endianness of the computer, only about
the endianness of the data to read.

Rewrite the function `toLittleEndian` and `fromLittleEndian` correctly.
Remove `toBigEndian` and `fromBigEndian` as data is always littleEndian.

As the data is always littleEndian, move the call `fromLittleEndian` in
the read function. Now `read` and `as` always return correct integer.

Fixes kiwix/kiwix-android#254