`find*` methods must also return inexact match.
`find*` methods are also use to find article starting by the query, so
we must return a valid iterator even if we have found a actual article
matching the query.

This API is very strange, we should change this. But other methods
assume this behavior, so keep it for now.