Remove Zint(stream) from libzim.
Zint (and its associated zintstream) was used to store information
in the embedded index (old format) or in the article's extra parameter.

None of them is used in libzim, so remove it.
Extra parameter is still available, but it is the responsibility of the
user code to store and read it as wanted.