Embedded gtest subproject in libzim source.
It seems that wrapdb.mesonbuild.org is not working.
It is a temporary problem, but it avoid us to compile libzim as
meson will try to use it to download extra gtest wrap.

This commit embed the gtest and meson wrap file directly in libzim
source instead of only declare it and let meson dowload source.

If meson found a correctly install gtest library already installed, it
will use it instead of compile its own version.

Embedding external project is rarely a good thing, but it should be ok as
gtest is only a "development dependency", it is not use by the installed
library.

The `subprojects/gtest` has been created the same way meson would have
made it :
- extract `https://github.com/google/googletest/archive/release-1.8.0.zip`
  (gtest source) in `subproject/gtest` directory
- extract `https://github.com/mesonbuild/gtest/archive/1.8.0.zip`
  (gtest wrap) in the same `subproject/gtest` directory.