Correctly protect temporary thread shared buffer.
The `bufferDirentZone` is a temporary buffer to store the raw
memory for a Dirent comming from the zim file.

We have to protect it when we use it in different threads else
dirent may be initialized with corrupt data.

Shame on me, I have defined the associated lock in the header but never
use :/

Fix kiwix/kiwix-tools#108.