Provide parse_html definition from HtmlParser in MyHtmlParser.
Recent versions of llvm are now compiled with '-Woverloaded-virtual`
option. (http://llvm.1065342.n5.nabble.com/Woverloaded-virtual-td5478.html)

This option raise a warning if a virtual method is overloaded in a chill
class. The compiler raise a warning because of the potential mistake of
overloading the method instead of overwriting it.

Fix #95

By explicitly bringing the definition of `parse_html` in `MyHtmlParser`,
we avoid the warning.