Improve ordering of suggestions search results.
By default xapian sort search result using relevance using the BM25
algorithm.

This alogrithm calculate the relevance using a "complex" calculation based
on the number of terms in each document, the number of terms in the
query, the number of documents, ...

This works pretty well with full documents but when we search in the
title namespace, this is not relevant at all.
Most titles are pretty short and the query is most of the time one term
only. In this case, all title are very close in number of term and all
contain only one term matching the query. Relevances are very closed and
most of the time irrelevant.

This commit sort the document using the Levenshtein distance of the
document titles to the query string. The search results are pretty more
relevant this way !

The levensthtein distance calculation algorithm is taken from wikibook
(https://en.wikibooks.org/wiki/Algorithm_Implementation/Strings/Levenshtein_distance#C++).