Use the correct type to offset and size.
Size were 32 bits and offset were 64 bits.
So we were using the offset_type everywhere we need a 64 bits integer even
if it was a size.

Now, size_type is a 64 bits, reuse the correct type everywhere.
We also use internal type `offset_t` and `zsize_t` to avoid implicit
conversion (and bugs).

As size_type change from 32 bits to 64 bits, we need to read a real
`uint32_t` in the zim when necessary.