Handle blob content bigger than 4go.
Offsets stored in a cluster were 32 bits. So we cannot handle article
bigger than 4Go.

This PR allow the `libzim` to store 64 bits offsets in a cluster.
To do so :
- Clusters "reader" and "writer" come in two flavors, one with 32 bits
  offsets and one with 64 bits offsets.
- The compression field became a bit field. The 4 lower bits are still
  used for compression (unchhanged). The 5th bit is use to store if
  the cluster uses 32 bits offsets (0) or 64 bits offsets (1).

Older versions of `libzim` will not be able to read extended cluster :
they will fail to read the compression field and raise a
zimFileFormatError.
So this change is a major change in the zim file format.
However, the new libzim will be able to handle both the old and the
new file format so we have to improve a bit the version handling.