Do not try to sort suggestions if too many matches.
The `set_sort_by_key` make the keyMaker being called once per relevant
documents.

Depending of the query string (ie. one or two chars), there will be
a lot of relevant documents (sometime more that 1 000 000) and the
calculation of the Levenshtein's distance can take a lot of time
(tests easily show around 5 seconds).

By trying to sort the matches only if they are not too many, we avoid
the request to take to many time.