Do not try to mmap the whole cluster.
If the (uncompressed) cluster is big, it can consumme a lot of memory.
Even more important, if the cluster is > 4GiB, we will mmap a potentially
wrong size on 32bits architecture as the maximum size is 4GiB.

So, now, we always return a (file) reader instead of a buffer reader on
the mmapped memory. As they are both reader, the API is not changed.
We will do the mmap only at real reading (when accessing the cluster header
or the blob).

At cost of :
 - doing several mmap instead of one.
 - Potentially have a duplication of the mmaps if the user code "ask"
   for several blobs (should we cache them ? assume that kernel will not
   really duplicate memory ?)

Also, we now don't have a big buffer handle by the cluster but a new buffer
per created blob. As buffer are ref count handled, `getPlobPtr` is invalid.
As soons as we return from the method, the created buffer is deleted and
so the returned address is no more valid.
The correct way to do is to get a blob and use the `Blob::data()`
(while keeping the blob "alive").