Ensure that a blob/buffer is never bigger than the addressable data.
If we are in 32 bits architecture, even if we can handle files bigger
than 4GiB, we still cannot address data bigger than 4GiB as the pointer
size is always 32bits.

If we try to get a blob of a content bigger than 4GiB, we will
return a empty Blob.

As we need to create a filereader from an existing std::FILE stream in test,
we also add corresponding constructor to internal structure.