Cast the value to OFFSET_TYPE.
`toLittleEndian` will use the type of the first argument (`o.v`) to
know the size of value and so, how many bytes to write.

But offset_t is always a 64bits integer and OFFSET_TYPE may be a 64bits or
a 32bits integer.

So `toLittleEndian` may write out of buffer.