Use the correct url when searching for xapian database.
Urls in zim file does not (and should not) start with a '/' and use
namespace 'X'.
The fact that xapian database has long url "Z//fulltextIndex/xapian" is
a mistake. Correct url should be "X/fulltextIndex/xapian".

To keep compatibility with already existing zim files, we must fallback
to the old (wrong) url if there is no new url.

Related to openzim/zimwriterfs#35