Revert "Search should return absolute url."
From discussion in PR kiwix/kiwix-lib#123 we agreed that absolute or
relative url is a non sense because url is more like a key than an real url.

The key should be what is store in the zim file itself (libzim give no
special meaning of to the url, expect the first char as namespace).

So, we must not add an extra `/` at beginning of the path.

This reverts commit 79b81e69198e7805ddf3147aed6b2fa63cafbfb9.