Avoid unnecessary copy of article content.
For binary article, the content is not modified and fully copied from
the file in the fs inside the zim file.
There is no need to read the file in memory, copy it in a cluster and
write the cluster into the tmp file. We can copy the file directly into
the tmp file.

We extend the zim::writer::Cluster to not directly store the final data,
but a list of "Data". The "Data" being either the real data (for in memory
Article) or a file name path (for file backed articles).

We also have to extend zim::writer::Article's interface to allow libzim
to know if the article has a file associated.

Fix #123