Move xapian indexer in libzim from zimwriterfs.
xapian database is a base libzim functionnality. The indexing part
should be in libzim.

This is mainly a recopy of code already in zimwriterfs in libzim.

The tools.cpp include the standard <dirent.h> header and it conflicts
with our "dirent.h".
To resolve the conflict, I've renamed dirent.h into _dirent.h.
As it is a private header, there is no problem :)

Fix #12