Better API for zimCreator.
The old API is based on an `ArticleSource`.  The main loop is in
`zimCreator` and uses the method `ArticleSource::getNextArticle()` to get
the next article to add.
This complexify a lot the implementation of `ArticleSource::getNextArticle`
because `ArticleSource` instance has to keep information about the current
state (iterators, which articles have been already added, ...)

With the new API, the main loop is external to `ZimCreator` and it has
to add article with `ZimCreator::addArticle()`.
User code is simpler because it just has to loop over the articles and
call `addArticle` at each step.