Better dirent and cluster handling.
One on the tricky thing is to track the cluster number of a dirent.
As we keep two different clusters opened at the same time (compressed and
uncompressed), we don't know the index of the cluster at the time we
add dirents to it.
So we need to keep a reference to "opened" dirents and once we close a
cluster, we update the index of dirents not closed.

With this change, we still have two opened clusters at the same time,
but instead of reusing a closed cluster as a new one to add new content,
we keep in memory (in a list) closed cluster and create a new cluster for
new content.

As clusters are not deleted nor reused, dirents can keep a pointer
directly on them. This pointer will never change, so we do not have to
keep a list of "opened" dirents. We will set the inde of the cluster when
we write the cluster, in the cluster itself. Dirents will use the pointer
to the cluster to retrieve the cluster index.

Once a cluster is writen, we don't need the data themselves, so we still
clean the cluster. (But it is not a reset anymore)

(This commit also move some dirent's methods specific to the writer into
zim::writer::Dirent instead of writer::Dirent)