Correctly sort titles' array.
`CompareTitle` will use the index to get the correct dirent in `dirents`.
As the indexes in `titleIdx` are "url indexes", the `dirents` must be
sort by url before sorting `titleIdx`.

The last time `dirents` was sorted was by aid in `resolveRedirectIndexes`,
so, `CompareTitle` were comparing title of wrong dirents.

Fix openzim/zimwriterfs#55