Add an option to avoid mmap (or copy) header index in memory.
By default we mmap/copy the ptrurl array in memory to have faster
lookup of where is the actual data.

But on low memory devices, mmap/copy reserve memory address and we may be
out of memory. By not copying the array in memory, we have more io calls
but don't use memory.