Do not use this_thread::sleep.
Mingw doesn't implement it. So, we should not use it.
I suppose that it was working before because mingw package for debian trusty
simply no provides a "thread" header.
We may face to include the native "thread" header.

Has sleep was only use for creator and test and we never try to launch
zimwriterfs on windows, we never face the problem.