Do not write the temporary article content is separated files.
This is somehow a potential unnecessary copy from the file containing
the article data in a temporary file.
However, with the new `mwoffliner` there is no file in the fs. Everything
is in memory. So writing the data is separated files is useless.
Even more, it is counterproductive as we need create a lot of temporary
files and we need to store their path.

Fix #224