Store pointer to dirent instead of dirent in the zimcreator vector.
This make Dirent manipulation faster as we have to swap the pointer and
not dirently the Dirent.

This also allow future code to reference dirent while we are creating
the zim as dirents will not be moved.

This change add memory leaks, but they will be fixed in future commits.

See #223