Index the title in a separated xapian database.
Previously we were indexing the title in a separated namespace(S) in the
same database.
Searching was made on this namespace, but the sorting was made by
computing rations using all the terms (including the fulltext ones)
and so the ratios were not good.
So we introduced the Leveinstein distance to mitigate this.
But the Levienstein distance need to access the title of each document and
this takes time.

By using a separated database for the title, the sorting use only
title terms and the ratios are far better. We don't need to sort
again with the Leveinstein distance and we are better.

If we use a "old" database, we still use the Leveinstein distance.
In case of multizim search, if one of the zims has a new database,
we stop using the Leveinstein distance.