Pass a shared_ptr to the creator instead of a reference [API break]
If we use a reference, it means that it is the calling code who is
responsible to delete the article so we cannot keep a pointer on the
article to index it later. And so we must do a copy of the content.

By passing a shared_ptr, the last "user" of the article will delete it.
So we can avoid copy of the data.