Correctly parse the query for suggestion on new database.
Fix kiwix/kiwix-tools#175

This is a bit tricky.

We used, at indexing, to stem the term using the STEM_ALL strategy.
This strategy stems all words and keep only the stemmed term as is.
At searching, we were doing the same when parsing the query.

But now, for title indexing (suggestion), we use the STEM_SOME strategy.
This strategy stems some words but also keep the unstemmed term.
The stem term is keep with the `Z` prefix.
If, at searching, we use the STEM_ALL strategy. We will search for
the stemmed term without `Z` prefix and so, we will never find it.
Using the STEM_SOME add the `Z` prefix to

As an example :

"marc antoine" is indexed using the STEM_SOME strategy some we have the
terms `marc`, `antoine` and `Zantoin`

If we search for "marc antoine" with STEM_ALL, we will search for
"marc AND (WILDCARD SYNONYM antoine OR antoin" and we will have results
(because marc and antoine are present)

But if we search for "antoine marc", we will search for
"antoin AND (WILDCARD SYNONYM marc OR marc)". As there is no term "antoin"
there will be no result.

If stem with STEM_SOME, the query will be
"Zantoin AND (WILDCARD SYNONYM marc OR Zmarc)". And we have results.