Reorder how we write zim to take advantage of not using a stream.
As we are not using a stream, we can seek in the file and know where
we are. So by writing cluster and dirents first, be can store where
we write them and so we can directly know the offsets without calculating
them.

The same for the header. Instead of calculating where the *PtrPos will be,
we write them first and set the value in the header. Then we write the
header when we know all the values.