Write all the clusters directly in the final zim file.
This is somehow the final step. Instead of writing clusters in temporary
files (in multi-threads), now we compress the clusters (in multi-threads)
and we write them (once at the time, in one thread) in the zim file.

However, zim format enforce that we write the zim header(known size) and
the mime-list(unknown size) before the other data. So we must start to
write the cluster at a guessed offset to keep space for the mime-list.

We reserve 1Kb. It allow us to have a bit less than 50 mime-list of 20
chars. It should be enough for our need.