Fix find function.
Commit fc8949b change the default order of `const_iterator`.
It breaks the `File::find` function as the find function find by url
and create a iterator using the default order.
As the order change, the iterator is broken.

This commit remove the default value of the `const_iterator`, so creating
a iterator need an explicit order.

The `File::begin` create the iterator using the "default" `ClusterIterator`
order.

This also add a method `File::beginByUrl` to allow user code to access
an iterator by url.

Fix kiwix/kiwix-lib#339