Correctly wait for the threads' termination before finishing zim creation.
We were doing the following in `Creator::finishZimCreation`:
- Wait for workers to finish their tasks
- Send a nullptr per worker as new tasks in the task queue
- [When a worker get a null tasks from the task queue,
   it quit its infinit loop (and so the thread ends).]
- We finish the zim creation and delete the creatorData (and so the
  task queue also).

What may happen is that the last steps of zim creation are really quick
and we delete the creatorData before all worker threads get the
null task from the queue. In this case, worker try to use the task queue
but it is deleted => SIGSEV.

By adding a `thread_join` we ensure that threads have totally end.
(Checking that all tasks have been made is not enough).