Correctly handle the fact that LZMA return a OK when needing more memory.
When we run out of out memory space, the first call to `lzma_code` will
return a `LZMA_OK` and following call will then return a `LZMA_BUF_ERROR`.

So we have to continue do `stream_run_encode` if the status is OK and
we are out of memory.

We do not check the `avail_in` as this can occur at `FINISH` steps and
in this case we are with no data.