PPA: Stop building for Ubuntu eoan, its EOL
Launchpad no longer accepts uploads for eoan.

While we're at it, restore the PPA as a build dependency for the package. I
had earlier removed it on the basis that zimlib didn't depend on any local
packages, but if we need to backport any extra dependencies (like I want to
for a newer debhelper), then we need to rely on the PPA. Also it's easier to
keep all the CI definitions the same anyways.