Fixed zim::FileCompound::locate() under MacOS
Under MacOS the implementation of `std::map::equal_range()` makes
assumptions about the properties of the key comparison function and
abuses the `std::map` requirement that it must contain unique keys. As a
result, when a map `m` is queried with an element `k` that is equivalent to
more than one keys present in `m`, `m.equal_range(k).first` may be different
from `m.lower_bound(k)` (the latter one returning the correct result).

This commit fixes the failure of the ZimFile.multipart unit-test under
MacOS (native_dyn).