More efficient concurrent access to cluster cache
Before this fix, if `FileImpl::getCluster()` was called from different
threads for the same cluster and the latter was not in the cache, all
threads would try to read and decompress the cluster (which is a waste
of CPU time). Then, multiple copies of the same cluster would be
returned from the concurrent calls to `FileImpl::getCluster()` (though
only one of them would end up in the cluster cache) and exist until
their `shared_ptr` handles are released.

This change makes sure that in the described scenario, only one of the
`FileImpl::getCluster()` calls reads and decompresses the cluster and
puts it in the cache. Any other threads in need of the same cluster are
blocked until it becomes available. The cache as a whole gets locked
only for the short time needed to access (find or create) the cluster
(or rather its placeholder, represented by a `std::future` object).