Make the Cluster use `IStreamReader`.
The cluster now internally use a `IStreamReader` to read the cluster data.
As we are reading the data sequentially, we have to store the previous
blob's data in a vector.

As we don't want to read/mmap the data of the blob we won't read, we don't
store a buffer but a reader.
Depending of the `IStreamReader`, the stored readers may be :
- A BufferReader (on the decompressed data). We have no choice here, we
  need to store the data if we don't want to decompress the data twice.
- A FileReader (on uncompressed data). This is a lightweight object with
  just a offset to the data to read. The data will be read only when we
  really want to access the blob.