DirentLookup got a lookup grid
DirentLookup now has a lookup-grid which is filled on initialization
with full paths (namespace + path) of equally spaced dirents. The first
stage of the lookup for the given namespace+path combination happens in
that in-memory lookup-grid (without access to actual dirents), providing
the possible range of the search item. The second stage then continues
as before by reading individual dirents and doing binary search on them.

What is missing in this initial implementation is the detection of
ranges that don't contain any dirents (for example, unused namespaces).

One notable change brought by this commit is that the result of a search
for an item in a non-existent namespace doesn't return an article index
of 0. Instead, the index of the first article from the next available
namespace is returned.