Handling of a ZIM file embedded in another file
Now it is possible to embed a ZIM file in another file and open it via
a file descriptor by also specifying the offset in the host file and the size of
of the ZIM file. This functionality is validated by a new unit-test
`ZimFile.openZIMFileEmbeddedInAnotherFile`.

The new test ZIM file `small.zim.embedded` was created by running

    test/data/create_test_zimfiles --only-derived-zimfiles