Fix creation of redirectDirent.
The `createDirent` method add the dirent for us using `addDirent`.
But `addDirent` test if the dirent is a redirection to add it
to `unresolvedDirents` or not.

So we need to call `dirent->setRedirect(nullptr)` **before** adding it.
And so we cannot use `createDirent`.

Let's create the dirent ourselves and configure it before adding it.