Handling of large MIME-type list
Unlike new-style ZIM archives created via libzim (where the first
cluster is placed exactly at offset of 1024 bytes and the dirent pointer
table is at the end of the archive), for old-style ZIM archives (where
the dirent pointer table is placed immediately after the MIME-type list
and before any clusters) there is no limit on the size of the MIME-type
list. Though 944 (1024-header_size) seemed to be a reasonable limit on
the size of the MIME-type list, it turned out that at least one
(old-style) ZIM file (stackoverflow.com_en_all_2019-02.zim) has been
created violating that assumption (but, to be fair, it is likely due to
a bug, since the MIME-type list in that file is a really weird one).
Whatever, we better be able to open any ZIM files conforming to the
spec, and the spec doesn't impose any limits on the size of the
MIME-type list.