Prevented an overflow in mimelist size estimation
This measure makes the following unit tests fail:
- ZimArchive.openingAnEmptyZimArchiveSucceeds
- ZimArchive.wrongChecksumInEmptyZimArchive

It can be avoided if the check uses < instead of <=, however the ZIM
format spec doesn't seem to allow a ZIM file (even an empty one) without
a MIME-type list section. Rather this change seems to have caught that
the empty ZIM archive in the unit test does not comply to the spec.