Use the DirentAccessor in DirentLookup.
While this change may seems trivial, it remove the dirent cache when
we lookup for dirent.

I THINK the DirentLookup is efficient enough and having another cache
may in fact be counter productive as we already search for dirent in a
subrange. But we need some testing to confirm this.