Allow reading of zero sized data on zero sized reader.
It is technically valid to read a zero sized at offset zero on a
zero sized reader.

The real assertion is that don't read past the end of the reader so
only `ASSERT(offset.v+size.v, <=, _size.v);` is needed.

This happen when reading a zim file with empty content and we don't
use mmap (on windows).
The cluster create a sub-reader of 0 size for the content and when
we try to read the 0 size content on the reader, we hit the ASSERT.
If we use mmap, we don't read from the reader, we create a mmap and
everything is fine.

The crashing assert is the one from FileReader but the logic is the
same for the BufferReader.

Fix kiwix/kiwix-lib#448