Make the cluster handle the extra 1 byte offset.
Initially, clusters were created from the clusterReader which was not
including the first info byte.
So all blob offsets were off by one byte regarding the cluster offsets
stored in the zim file.

But now we are creating the cluster directly from the its offsets.
So it is better to move this extra offset in cluster directly.