Fix default indexing of data.
The "api" of htmlParser is a bit surprising, it always end with
a exception, even if everything is ok :/

Discarding the exception all the time far from being perfect, but
except changing the parser it is not so easy.

As this is a default indexing and we want to creator implement their
own (specialized) parser (and because it was already working this way)
let's go with this.