Move the queryParser in the InternalDataBase.
As the QueryParser is only dependent of the information in the database
we can directly create it as we create the database.

We don't need a specific method to configure a QueryParser every time.

As m_language and m_stopwords were used only to configure the
queryParser, we don't need to store them in the database.