Fix searching of main entry in new zim files.
The "Main Page" entry is a redirection pointing to the real
main page.
But now, the redirection itself is not searchable, so we cannot
search for "Main Page". We have to use the real title of the target item.

On wikibooks_be_all_nopic_2017-02, the new zim file now have a
titleIndex so we run the "search test" in `checkEquivalence`.
But wikibooks_be_all_nopic_2017-02 contains two items with the title of
the mainPage :
- The (real) entry itself.
- A redirection pointing to the real entry.

So, we need to also resolve the redirection of the search result to check
that path correspond.