Add a new custom target allowing to download the test data from meson.
The target is not run by default.
It allow user to download the data itself if he wish.
To download the data using meson, one must run `ninja download_test_data`
once configured.

The path to the data dir must be passed to meson using the
`test_data_dir` option (`-Dtest_data_dir=/foo/test_data`).
If the option is not passed, default to a subdirectory inside the build
directory.