Fix build on Windows.
On Appveyor, cl.exe include the wrong "tools.h".
It works on linux/github because meson add the "src" directory to the
include path.
On Appveyor there is probably another "tools.h" in the include directories
and we include the wrong one.

The correct way to include "tools.h" is to use relative path.