Lower zstd compression level.
The max level (22) set a huge windowSize (128Mb).
It if far too much and we don't even need it (most of our compressed
content is about 1Mb).

By setting a level of 19, the windowSize is 8Mb and
the compression ratio is relativly equivalent.
From lzbench https://github.com/inikep/lzbench the ratio is
24.88 for level 22 and 25.33 for level 18.

See kiwix/libkiwix#515