Correctly handle exception in concurentCache.
We must handle exception raised in the callable in `getOrPut`.

If not, we would have a broken promise in the cache.
It is better to have no promise in the cache at all if we fail to
create the value. (The concurent cache user as no notion of promise).

And sometime, the creation of the value is temporarly failing.
For exemple, in kiwix/libkiwix#515, if some memory is free, we would
be able to create a cluster even if we failed the last time.