Do not cache faulty cluster.
Some clusters are problematic on 32bits as they are reserving 128Mb buffer.
(Even if they don't use the memory).

This commit try to remove the cluster from the cache to mitigate this
issue.

The cluster is removed from cache if :
- We are in 32bit,
- and we are using zstd compression.
- and the zim file is "too old".

The "too old" detection is made using the major.minor version.
This is not perfect as we will generate "good" zim file with version
5.0 and so we may drop good clusters but it what is simpler.
And will be less and less "positive" zim files (false or true) with time.