Introduce SuggestionSearcher to search API
SuggestionSearcher mimics the old Searcher in every way except that it
does not ONLY depend on an internalDb to generate a search. This is done
to facilitate suggestions for archives without title index.

As a part of this new Suggestion API, we no longer need to store
suggestion mode in query and no longer need a separate pointer to
internalSuggestionDb, initDatabase() is also adjusted to this new
scheme.