Fix Archive::iterator copy constructor
Currently the copy constructor uses the dereference of the other.m_entry
pointer of other iterator to set the new m_entry. But the m_entry
pointer is only updated after calling the `operator*()` and hence there
is a potential risk of dereferencing a nullptr. To mitigate the risk, we
should directly update the m_entry by dereferencing the other iterator.