Introduce TinyString object to store strings efficiently.
Sizeof `std::string` is 32B. It is too big for use.
TinyString only need a size(8B) and a pointer(8B).
So we store string with half less memory.

As we use 3 string we reduce our size of 48B
at the cost of recreate the string when we need to use it.
The new dirent size is 96B instead of 144B.