Break reference loop in `Item`.
The basic idea was to store the created indexData's `shared_ptr` to
avoid creating it twice.

But the contentProvider may store a reference to the item, especially
when it reference content stored in the item (as it is the case for
`SharedStringContentProvider`.

This commit remove any storage of the created indexData.
This break the loop but at the cost of creating twice the indexData.