Use only one handler for title listings.
As we have now "free space" to store boolean in the dirent, we can store
if it is a front article in it.
So we don't have to keep a separated vector of Dirent pointer in
`TitleListingHandlerV1`. We can have only one handler, generating two
items. The second being the v1 listing which contains only front articles.

This allow us to earn 8 bytes (a pointer) per front article.