Switch default_op for query parser to OP_AND
The default_op for the query parser is already being set during the
initialization of database. The reconfiguration in
SuggestionDataBase::parseQuery() is redundant and wrong. We should
parse suggestion queries only  with OP_AND and OP_PHRASE to ensure
relevant suggestions.

This bad-configuration bug came from a buggy separation of suggestions
and FT search and it didn't cause any issue till now since we were
parsing the first query with the default OP set during initialization.
But it surfaced during implementation of caching since we were now using
the bad-reconfiguration.