Protect parseQuery with a mutex lock and reset stemming
For proper parsing of queries, after first parsing with default
STEM_SOME strategy, we reconfigure the stemming strategy to
STEM_NONE for phrase search. To reuse the searcher properly we need to
reset the stemming strategy back to STEM_SOME. This can lead to some
issues in case of multithreading.

To prevent such weird behavior, we must protect the parsing with a mutex
lock.