Do not try to dereference end() iterator in the back of our user.
The (previous) iterator copy constructor implied a dereferencing of
`other` iterator.
So when, we copy the end iterator, we try to dereference it, which may
throw a `out_of_range` exception if the end iterator's idx is after the
valid indexes.

Now we do a copy of the Entry only if `other` iterator has an existing
Entry.