findxByTitle now search for idx a start of identical titles range.
`findxByTitle` is used to search for exact match (and so we don't care
about this commit) but also in `Archive::findByTitle` to find a range of
indexes for whom the entries start by the title.

But it is not mandatory that titles are unique. So we must ensure that,
if we've found a exact match, we return the index of the first entry with
the exact match. We cannot simply stop to search when we have found one.

As the entries are title ordered, we can simply go backward once we've
found a exact matching to be sure to return a index at the beginning of the
range.