Simplify findxByTitle implementation.
The `else` branch in the loop was a early exit if we have found a
exact match.
But we don't want to just find a exact match, we want to find the beginning
of the exact match range.
The easier is to not do a early exit at all and reduce the bounds the
smaller possible around the range beginning.
And then "resolve" the final bounds to check for exact match or not.