New unit-test ZimCreator.interruptedZimCreation
The new unit test tries to provide some evidence that interrupted
creation of a ZIM file (when `zim::Creator` is destroyed without
`zim::Creator::finishZimCreation()` having been called) doesn't end up
similar to openzim/zim-tools#289.

So far this is not the case, that's why the new unit-test fails with a
crash.