Correctly mark removed dirent as removed.
At creation, we have to handle cases where the user add Entry to the
zim file but we remove it later.
Their are two cases:
- When user add a Redirection, we need to accept it even if it points to
  a non existing entry (as the entry may be added later).
  At the end of the process, we loop on all redirection and we removed
  dandling ones.
- When user add two Entries with the same path. If the first added entry
  is a redirection and the second is an item, we remove the redirection
  and accept the item.

But we handle the added dirent in specific handle (indexer, listing) as
soon as the entry is added.

So we must mark the removed dirent and so, the specific handlers will be
able to exclude removed dirents at the end of the process.

We were marking the dirent as removed in `resolveRedirectIndexes`
(the first use case) but forget in the `addDirent` (the second use case)
This is now fixed.