Throw an exception if the cannot add an entry to the creator.
If user want to add two entries with the same path, we cannot do it.

This is detected at pretty low level (but early) on our side.
- We must stop the adding process when we detect this.
- We must warn the user (better than print on stderr).

Exception are made for this, even if the user has to handle explicitly it.

There is a small memory leak here as a dirent is reserved for second
entry to add and it is never release (but it is memory freed correctly at
the end of the process). This is a small memory leak and it should be a
somehow rare case (it is not normal to add two entries with the same path)