Support for branch name or tag name in checkout
Fixes openzim/libzim#450

The new github.ref_name should automatically populate with either
the branch name or tag name depending on what kicked this off.
Using this eliminates the need for the step to generate this name.

Attempted to move entirely to using Github's `actions/checkout` but
wasn't able to because the `kiwix_build_ci` images have a built-in
`runner` user. The checkout attempts to use the `/__w/` directory that
has been mounted from the host machine, but `runner` can't write to it.