Added Entry::getRedirectEntryIndex()
`Entry::getRedirectEntryIndex()` is functionally equivalent to
`Entry::getRedirectEntry().getIndex()`. However, the latter involves
access to the dirent of the target entry. When only the index of the
target entry is of interest, unnecessarily loading the target dirent may
turn sequential disk IO (when iterating entries by their native order)
to somewhat random disk IO. The new method is devoid of that drawback.