Remove transaction system from xapian indexer.
Transactions are used to regroup modifications together and be sure that
they are either written all together in the database or not at all.

But here we don't care about coherence of the database. Either we succeed
to write index all documents or we have fail all the process.
And we never read the database will we are creating it.

On top of that, transaction probably prevent xapian to automatically
commit the modifications and so we are keeping everything in memory and
commit in fs only at the end of the process.

By removing transactions, we let xapian handle commits "correctly".
By default xapian will commit every 10000 documents.
This can be changed with the env var XAPIAN_FLUSH_THRESHOLD
See https://xapian.org/docs/apidoc/html/classXapian_1_1WritableDatabase.html#acbea2163142de795024880a7123bc693